Update alphafold.md
Replace link text 'here' by something more meaningful.

See also:

https://archive.org/embed/designingwebusab00niel
Designing Web usability, Nielsen, Jakob, 2000
https://archive.org/details/designingwebusab00niel/page/54/mode/2up?q=blue+bee

The oldest web design rule is to avoid using "Click Here" as the anchor text
for a hypertext link. There are two reasons for this rule. First, only
mouse-using visitors do in fact click, whereas disabled users or users with a
touchscreen or other alternative device don't click. Second, the words "Click"
and "Here" are hardly information-carrying and, as such, should not be used as a
design element that attracts the user's attention. Instead of saying  

For background information on the blue-nosed honeybee, click [here](dummy).  

It is better to say  

We have additional background information about [the blue-nosed honeybee](dummy).